requirements.txt
streamlit
pandas
scikit-learn
requests
textblob
python-dotenv